refactor: delete proven duplicate code
